+++ doc/whats_new.rst
@@ -90,6 +90,8 @@ Bug
 
 - Fix bug in :func:`mne.preprocessing.maxwell_filter` where homogeneous fields were not removed for CTF systems by `Eric Larson`_
 
+- Fix bug in writing ``raw.annotations`` where empty annotations could not be written to disk, by `Eric Larson`_
+
 - Fix support for writing FIF files with acquisition skips by using empty buffers rather than writing zeros by `Eric Larson`_
 
 API